Add basic Dockerfile
The Dockerfile runs the 'analyze' command.
Usage instructions:
docker build -t ac_extractor .
docker run ac_extractor [arguments?]

Currently the run fails because it can't find MusicExtractor and FreesoundExtractor (are these in the essentia image?)
Then we need to think how to handle file input and ouput. The docker container needs to be able to read input and write to output.